[build] Turn dev-tools into an optional feature
Introduce a feature flag "devtools" for toggling these tools on and off
