+++ Cargo.toml
@@ -31,13 +31,24 @@ safemem = "0.3"
 tempfile = "3.3"
 threadpool = "1.8"
 thiserror = "1.0"
-tui = "0.14"
-termion = "1.5"
+tui = { version = "0.14", optional = true }
+termion = { version = "1.5", optional = true }
 
 [dev-dependencies]
 json = "0.12"
 quickcheck = "0.9"
 quickcheck_macros = "0.9"
 
+[features]
+devtools = ["tui", "termion"]
+
 [profile.release]
 debug = true
+
+[[bin]]
+name = "cache_generate_metadata"
+required-features = ["devtools"]
+
+[[bin]]
+name = "thin_explore"
+required-features = ["devtools"]
